+++ java-design-annotation/src/main/java/nbbrd/design/Demo.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.design;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -27,6 +28,7 @@ import java.lang.annotation.Target;
  */
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.SOURCE)
+@Documented
 public @interface Demo {
 
 }