+++ java-design-annotation/src/main/java/nbbrd/design/RepresentableAs.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.design;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -30,6 +31,7 @@ import java.lang.annotation.Target;
  */
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.CLASS)
+@Documented
 public @interface RepresentableAs {
 
     Class<?> value();