+++ java-design-annotation/src/main/java/nbbrd/design/SealedType.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.design;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -29,6 +30,7 @@ import java.lang.annotation.Target;
  */
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.SOURCE)
+@Documented
 public @interface SealedType {
 
     Class<?>[] value();