+++ java-design-annotation/src/main/java/nbbrd/design/VisibleForTesting.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.design;
 
+import java.lang.annotation.Documented;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 
@@ -24,5 +25,6 @@ import java.lang.annotation.RetentionPolicy;
  * relaxed to make the code testable.
  */
 @Retention(RetentionPolicy.SOURCE)
+@Documented
 public @interface VisibleForTesting {
 }