+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix SystemDependent annotation target
+- Fix missing Documented annotation
 
 ## [1.10.0] - 2026-04-14